+++ plugins/packages/langfuse-importer/pyproject.toml
@@ -22,7 +22,10 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 keywords = ["ai-agents", "kitaru", "langfuse", "observability", "traces"]
-dependencies = ["kitaru>=0.22.0"]
+dependencies = [
+    "kitaru>=0.23.0",
+    "langfuse>=3.0.0,<4",
+]
 
 [project.urls]
 Homepage = "https://kitaru.ai"
