+++ plugins/packages/langfuse-importer/src/kitaru_langfuse_importer/__init__.py
@@ -1,1 +1,7 @@
 """Langfuse trace importer for Kitaru."""
+
+from .adapter import LangfuseAdapter
+
+__all__ = [
+    "LangfuseAdapter",
+]
