+++ plugins/pyproject.toml
@@ -8,7 +8,6 @@ members = ["packages/*"]
 kitaru = { path = "..", editable = true }
 kitaru-braintrust = { workspace = true }
 kitaru-braintrust-importer = { workspace = true }
-kitaru-langfuse = { workspace = true }
 kitaru-langfuse-importer = { workspace = true }
 kitaru-langgraph = { workspace = true }
 kitaru-langsmith = { workspace = true }
@@ -24,7 +23,7 @@ kitaru-pydantic-ai = { workspace = true }
 dev = [
     "kitaru[server]",
     "kitaru-braintrust",
-    "kitaru-langfuse",
+    "kitaru-langfuse-importer",
     "kitaru-langgraph[deepagents]",
     "kitaru-langsmith",
     "kitaru-logfire",
@@ -56,7 +55,6 @@ known-first-party = [
     "kitaru_braintrust",
     "kitaru_braintrust_importer",
     "kitaru_jsonl_importer",
-    "kitaru_langfuse",
     "kitaru_langfuse_importer",
     "kitaru_langgraph",
     "kitaru_logfire",
