+++ tests/scripts/test_release_units.py
@@ -29,7 +29,6 @@ EXPECTED_UNITS = {
     "braintrust-importer": "kitaru-braintrust-importer",
     "evaluator": "kitaru-evaluator",
     "jsonl-importer": "kitaru-jsonl-importer",
-    "langfuse": "kitaru-langfuse",
     "langfuse-importer": "kitaru-langfuse-importer",
     "langgraph": "kitaru-langgraph",
     "logfire": "kitaru-logfire",
@@ -121,7 +120,7 @@ def core_release_repo(
     return tmp_path, version
 
 
-def test_inventory_describes_core_and_fifteen_plugin_distributions() -> None:
+def test_inventory_describes_core_and_fourteen_plugin_distributions() -> None:
     inventory = load_inventory()
 
     assert {unit.slug: unit.distribution for unit in inventory.units} == EXPECTED_UNITS
@@ -645,7 +644,7 @@ def test_text_and_json_outputs_contain_the_same_unit_identities() -> None:
     assert all(unit.distribution in text_output for unit in inventory.units)
 
 
-def test_plugin_matrix_is_generated_from_the_fifteen_plugin_units() -> None:
+def test_plugin_matrix_is_generated_from_the_fourteen_plugin_units() -> None:
     matrix = build_plugin_matrix(load_inventory())
 
     assert matrix == {
@@ -835,7 +834,7 @@ def _run_cli(*arguments: str) -> subprocess.CompletedProcess[str]:
     [
         (["list"], "SLUG\tDISTRIBUTION\tVERSION\tDEFAULT\tTAG"),
         (["resolve", "--unit", "kitaru"], "python/kitaru/v"),
-        (["validate"], "Validated 16 release units."),
+        (["validate"], "Validated 15 release units."),
         (
             [
                 "propose-core-version",
